src/uib_serial_sim/requirements.txt
crccheck==1.0
dataclasses==0.8
pyserial==3.5
